Add output and filter option to gp mode
Different output modi for transcription are now available. An
option for a fixed replacement on IPA side has been added as well:
It is now possible to replace undesired IPA symbols (e.g. the
incorrect affricate-consonant representation with two symbols and
a connecting arc frequently used in the German wiktionary).

Fixes #12